fix: named import is failing with esm node
